+++ packages/data/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @apex-stack/data
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @apex-stack/core@0.15.0
+
 ## 0.4.0
 
 ### Minor Changes