+++ playground/showcase/CHANGELOG.md
@@ -1,5 +1,13 @@
 # showcase
 
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @apex-stack/core@0.15.0
+  - @apex-stack/data@0.4.1
+
 ## 0.0.36
 
 ### Patch Changes